requirements.txt
python-telegram-bot[job-queue]==21.6
APScheduler==3.10.4
pytz==2024.1
